Refact : SeacrhResult 리팩토링
TitleResult로 컴포넌트 명 변경
로딩중 조건변경
무한 스크롤 로직 제거
props 추가
