+++ finda2.0/src/components/Result/TitleResult/Index.style.ts
