+++ finda2.0/src/components/Result/TitleResult/type.ts
